+++ packages/tools/src/drawingSvg/getSvgDrawingHelper.ts
@@ -38,7 +38,10 @@ function getSvgDrawingHelper(element: HTMLDivElement): SVGDrawingHelper {
 function _getSvgLayer(element) {
   const viewportElement = `.${VIEWPORT_ELEMENT}`;
   const internalDivElement = element.querySelector(viewportElement);
-  const svgLayer = internalDivElement.querySelector('.svg-layer');
+
+  // Using :scope to make sure the right svg layer is selected otherwise it
+  // may select one from a nested viewport (eg: AdvancedMagnifyTool).
+  const svgLayer = internalDivElement.querySelector(':scope > .svg-layer');
 
   return svgLayer;
 }