+++ packages/tools/src/utilities/segmentation/utilities.ts
@@ -10,7 +10,10 @@ export type ThresholdInformation = {
   upper: number;
 };
 
-export default function getBrushToolInstances(toolGroupId) {
+export default function getBrushToolInstances(
+  toolGroupId: string,
+  toolName?: string
+) {
   const toolGroup = getToolGroup(toolGroupId);
 
   if (toolGroup === undefined) {
@@ -21,6 +24,10 @@ export default function getBrushToolInstances(toolGroupId) {
 
   if (!Object.keys(toolInstances).length) {
     return;
+  }
+
+  if (toolName && toolInstances[toolName]) {
+    return [toolInstances[toolName]];
   }
 
   // For each tool that has BrushTool as base class, set the brush size.